// frontend/src/generated/models/OtaDevice.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tempmon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface OtaDevice
 */
export interface OtaDevice {
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    ipAddress: string;
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    deviceName: string;
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    version: string;
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    deviceBoard: string;
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    chipModel: string;
    /**
     *
     * @type {string}
     * @memberof OtaDevice
     */
    wifiRssi: string;
    /**
     *
     * @type {number}
     * @memberof OtaDevice
     */
    lastSeen: number;
}

/**
 * Check if a given object implements the OtaDevice interface.
 */
export function instanceOfOtaDevice(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "ipAddress" in value;
    isInstance = isInstance && "deviceName" in value;
    isInstance = isInstance && "version" in value;
    isInstance = isInstance && "deviceBoard" in value;
    isInstance = isInstance && "chipModel" in value;
    isInstance = isInstance && "wifiRssi" in value;
    isInstance = isInstance && "lastSeen" in value;

    return isInstance;
}

export function OtaDeviceFromJSON(json: any): OtaDevice {
    return OtaDeviceFromJSONTyped(json, false);
}

export function OtaDeviceFromJSONTyped(json: any, ignoreDiscriminator: boolean): OtaDevice {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'ipAddress': json['ipAddress'],
        'deviceName': json['deviceName'],
        'version': json['version'],
        'deviceBoard': json['deviceBoard'],
        'chipModel': json['chipModel'],
        'wifiRssi': json['wifiRssi'],
        'lastSeen': json['lastSeen'],
    };
}

export function OtaDeviceToJSON(value?: OtaDevice | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'ipAddress': value.ipAddress,
        'deviceName': value.deviceName,
        'version': value.version,
        'deviceBoard': value.deviceBoard,
        'chipModel': value.chipModel,
        'wifiRssi': value.wifiRssi,
        'lastSeen': value.lastSeen,
    };
}